+++ gym-tracker-ui/src/pages/WorkoutPage.vue
