+++ gym-tracker-ui/src/router.ts
@@ -3,7 +3,7 @@ import HomePage from './pages/HomePage.vue'
 import LoginPage from './pages/LoginPage.vue'
 import NotFoundPage from './pages/NotFoundPage.vue'
 import Layout from './components/Layout.vue'
-import WorkoutsPage from './pages/WorkoutsPage.vue'
+import WorkoutPage from './pages/WorkoutPage.vue'
 const routes = [
     { path: '/', redirect: { name: 'home' }, name: 'root' },
     { path: '/login', component: LoginPage, name: 'login' },
@@ -13,7 +13,7 @@ const routes = [
         redirect: { name: 'home' },
         children: [
             { path: 'home', component: HomePage, name: 'home' },
-            { path: 'workouts', component: WorkoutsPage, name: 'workouts' },
+            { path: 'workout', component: WorkoutPage, name: 'workout' },
         ],
     },
     { path: '/:pathMatch(.*)', component: NotFoundPage },
